requirements.txt
# Core dependencies
aiohttp>=3.8.0
discord.py>=2.0.0
python-dotenv>=0.19.0
python-dateutil>=2.8.0
requests>=2.25.0

# AI/ML
anthropic>=0.3.0
numpy>=1.20.0
opencv-python-headless>=4.5.0
openai
google-generativeai>=0.3.0
google-genai

# File handling
aiofiles>=0.8.0
pillow>=8.0.0
moviepy>=1.0.3
imageio>=2.31.0
imageio-ffmpeg>=0.4.8  # Bundles ffmpeg binary - more reliable in containers
av>=10.0.0  # PyAV plugin for imageio video reading
yt-dlp>=2024.0.0  # External media extraction for TopicEditor

# Scheduling
apscheduler>=3.9.0

# APIs & Services
praw>=7.7.0
supabase>=1.0.0
tweepy>=4.0.0

# Database
psycopg2-binary>=2.9.0
hypothesis>=6.0.0

# Solana
solana>=0.34.0
solders>=0.21.0
base58>=2.1.0
